Extract `SortableHeader` component
