+++ static/app/views/starfish/modules/databaseModule/databaseTableView.tsx
@@ -10,11 +10,11 @@ import {CHART_PALETTE} from 'sentry/constants/chartPalette';
 import {space} from 'sentry/styles/space';
 import {Series} from 'sentry/types/echarts';
 import {getDuration} from 'sentry/utils/formatters';
+import {SortableHeader} from 'sentry/views/starfish/components/sortableHeader';
 import Sparkline, {
   generateHorizontalLine,
 } from 'sentry/views/starfish/components/sparkline';
 import {Sort} from 'sentry/views/starfish/modules/databaseModule';
-import {SortableHeader} from 'sentry/views/starfish/modules/databaseModule/panel/queryTransactionTable';
 
 type Props = {
   isDataLoading: boolean;
